:zap: Experiment some more with group/collection displays
The Raindrop API doesn't seem to be that coherent when it comes to handling
this sort of thing, and there seems to be more than one method of getting at
collections and their contents.

Here I'm making a start on a widget that will handle this. I'm not sure I
like this approach yet but it's a useful riff for now.

The one problem I have here is that I'm not diving into the whole hierarchy,
and also I'm going via the User and then getting all the collections anyway.

I feel there should be a better API for this. In playing around with this
I've also noticed that the API documentation doesn't actually quite match
the actual API; there's a bunch of stuff missing from the documentation, it
seems.