:hammer: Rework the data and widget interaction to start downloading
This is a bit of a rework of the code layout, plus work to do the download,
to get to a state where I can detect if there is the need to download data
from the server. It also changes how the navigation pane gets populated.

At the moment it still doesn't display anything in the main pane, that comes
soon.