:hammer: Make the navigation one single widget
Initially I was going to make the navigation sidebar a collection of
different OptionList-based widgets, but that wasn't feeling right. So now my
idea is to try and make it a single list, which will include all special
collections, all user-defined collections (within their groups), and
eventually tags and the like.

In this I'm also leaning into the disabled options to use them as headings.