:lipstick: Make the Textual command palette look less unfinished
It seems that, at some point, someone changed the command palette so that
it's the full width of the screen; this makes it look like unfinished and
forgotten about garbage. I don't want my app to look like it contains
unfinished and forgotten-about garbage.

Making the command palette a wee bit more narrow than the screen makes it
look like it's actually floating above the underlying screen.