:lipstick: Ensure we never get a flash of Tags scrollbar
I want the list of tags in the details view to always be exactly as high as
it needs to be, and so I don't ever need to see the vertical scrollbar. On
the other hand I was seeing it appear for a brief moment before it went
away, at times; presumably something to do with how Textual draws things.

This makes it so that we never see that flash.