:hammer: Add default values for all Raindrop fields
I'm about to move into a state where I want to make a Raindrop object, but
with partial data, because I want to use it to post it back to the server.
Rather than have to declare a default value everywhere all the time, here I
ensure the Raindrop class itself does that.