:sparkles: Add a as_json property to the Raindrop class
This is still a little tentative as it's mostly here to serve the API when
passing data back to the server; rather than being something tat can be used
to store the data, etc.

Later it might make sense to make it so that it simply returns self.raw if
that isn't empty, otherwise construct the data.