:hammer: Change how I decide if a fresh download needs to happen
Until now it's always been a case of has the last action on the server
happened later than the last download time, to the deepest fraction of a
second.

I now think it's possible for raindrop.io to perform (and record) an action
but not make any changes; an example would be for pro users where it might
be seeing if a new link is broken, recording the last action time as when it
did that, but not actually update because the link wasn't broken.

So here I'm changing this to use the last update time for now, and also
giving it a little bit of wiggle room for the time difference; this is
working on the assumption there could be a fraction of a second difference.

Note to self: I should also ensure I always make this comparison in a
timezone-safe way. I think I am right now, but I need to be 100% sure.