:sparkles: Add the ability to set the raindrop view's highlighted raindrop
Here you can set it to None (so nothing highlighted -- this really just does
what setting highlight to does) or to a specific Raindrop object. If it
doesn't exist then this is a no-op.

The idea here being that I can see the highlighted raindrop iff it's in the
current view, otherwise the current view is undisturbed.