:hammer: Don't force a reload of the new raindrop dialog on error
I think it makes more sense to just keep the newly-input data as a draft and
let the user discover that it's still there if they try and create the
raindrop again. Whatever error happens could be something they want to act
upon, so it doesn't make sense to force them into a dialog they might not
want to be in right now.