:hammer: Always edit a raindrop object
Either start out with the one passed in, or a freshly-created one, and then
patch a clone of it with the new values before passing it back. This way I
can safely add more and more support for the content without anything being
lost when it comes to the edit operation.