:sparkles: Add the ability to edit a raindrop
This needs a wee bit more work, but it's the basic layout of the code. The
one big issue here (and with adding a raindrop too really) is that, once
done, we force going back to the "All" view. I don't want this.

It's good enough to get all the other parts in place, but I want adds and
edits to not disturb the state and UI more than is necessary.