:green_heart: Try testing on macOS as well as GNU/Linux
I would like to test on Windows too, but it doesn't have a native make and
faffing with the config to ensure there is one seems like a lot of
bother (perhaps later). The alternative is to use the appropriate rye
command right in the config but that sort of defeats the purpose of having a
Makefile.