:lipstick: Improve the cosmetics of the main focused panels
The main difference here being that I downplay the scrollbars (they always
stand out way too much for applications that are intended to be
keyboard-first) and also drop the background-tint nonsense in Textual's
builtin OptionList, as it then puts it in a different styling category as
the scrollbars (as best as I can see there's no direct support for a
background-tint for them).

Overall I feel this commit makes things feel more smoother and consistent.