:bug: Fix the doubled-up help text in the raindrops details panel
The help was doubled up in the tags widget as well as in the panel itself;
this had originally been done to make things work fine with Textual's own
help panel. However, when I realised it wasn't a good approach to showing
help for keyboard users I switched to my own help dialog, but forgot to undo
this workaround.

Fixes #39