:card_index: Help mypy be less confused
It was insisting that widget_type was Widget, and so there was no update
attribute; almost like it was taking the union of Label and Markdown and
working with the nearest common ancestor.

Meanwhile pyright didn't seem to care one jot and understood what I meant.

Casting it seems to help.