:hammer: Rework the way that filtering works
This adds no substantive change right now, but it does set me up for any
future changes where I want to add more flexible searching and filtering.
Written with half an eye on #89.

At the moment this just reimplements searching for tags and free text
anywhere in a field, but it'll make it easier to add things like only
searching in a specific field, etc.