:zap: Exclude unknown collections when getting a group's collections
While I can't recreate it myself, looking at #104 it would appear that
Raindrop can be got into a state where the API will report that a group has
collections within it that don't actually exist any more. This commit seeks
to ignore such collection IDs.

So, with this change, a collection ID that isn't known to us is treated as
if it doesn't exist.